 requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
appium-python-client==5.1.2
asynch==0.3.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
ciso8601==2.3.3
click==8.2.1
clickhouse-driver==0.2.9
clickhouse-sqlalchemy==0.3.2
colorama==0.4.6
colorlog==6.9.0
cryptography==45.0.6
dnspython==2.7.0
execnet==2.1.1
faker==37.5.3
flask==3.1.2
greenlet==3.2.4
h11==0.16.0
idna==3.10
iniconfig==2.1.0
invoke==2.2.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonpath==0.82.2
jsonpath-ng==1.7.0
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
leb128==1.0.8
lz4==4.4.4
markupsafe==3.0.2
numpy==2.2.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
paramiko==4.0.0
pluggy==1.6.0
ply==3.11
prettytable==3.16.0
pycparser==2.22
pyecharts==2.0.8
pygments==2.19.2
pymongo==4.14.1
pymysql==1.1.1
pynacl==1.5.0
pyqt5==5.15.11
pyqt5-qt5==5.15.2
pyqt5-sip==12.17.0
pysocks==1.7.1
pytest==8.4.2
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==15.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.2
selenium==4.34.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.43
trio==0.30.0
trio-websocket==0.12.2
typing-extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.13
websocket-client==1.8.0
werkzeug==3.1.3
wsproto==1.2.0
zstd==1.5.7.2
